Update main.py
You're getting a “container failed to start / didn’t listen on PORT” because the app likely crashed on import. I spot a bug in the materialize_cli_config function I sent: there’s an extra } inside the f-string, which will throw a SyntaxError at startup and make the container never bind to 8080.